refactor: move parse_header to own file
